% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG_Benthic.R
\name{BCG_Benthic}
\alias{BCG_Benthic}
\title{Function to call the Benthic BCG model}
\usage{
BCG_Benthic()
}
\value{
Calls the app for using the Benthic model
}
\description{
Function to call the Benthic BCG model
}
