% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install_Capture.R
\name{install_capture}
\alias{install_capture}
\title{Install Capture}
\usage{
install_capture()
}
\value{
Installs the package \code{capture} from GitHub
}
\description{
Install Capture
}
\keyword{internal}
