% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG_Format.R
\name{BCG_Formatter}
\alias{BCG_Formatter}
\title{Function to call the BCG do-it-all formatter}
\usage{
BCG_Formatter()
}
\value{
Calls the app for using the BCG Formatter
}
\description{
Function to call the BCG do-it-all formatter
}
